split star transpiler into placer and router
