+++ tests/test_transpiler_router.py
@@ -10,8 +10,21 @@ from qibo.transpiler.pipeline import (
     assert_circuit_equivalence,
     restrict_connectivity_qubits,
 )
-from qibo.transpiler.placer import Custom, Random, Subgraph, Trivial, assert_placement
-from qibo.transpiler.router import CircuitMap, Sabre, ShortestPaths, assert_connectivity
+from qibo.transpiler.placer import (
+    Custom,
+    Random,
+    StarConnectivityPlacer,
+    Subgraph,
+    Trivial,
+    assert_placement,
+)
+from qibo.transpiler.router import (
+    CircuitMap,
+    Sabre,
+    ShortestPaths,
+    StarConnectivityRouter,
+    assert_connectivity,
+)
 
 
 def star_connectivity():
@@ -331,3 +344,60 @@ def test_restrict_qubits(router_algorithm):
     )
     assert_connectivity(restricted_connectivity, routed_circ)
     assert_placement(routed_circ, final_layout, connectivity=restricted_connectivity)
+
+
+def test_star_error_multi_qubit():
+    circuit = Circuit(3)
+    circuit.add(gates.TOFFOLI(0, 1, 2))
+    transpiler = StarConnectivityRouter(middle_qubit=2)
+    with pytest.raises(ConnectivityError):
+        transpiled, hardware_qubits = transpiler(circuit)
+
+
+@pytest.mark.parametrize("nqubits", [1, 2, 3, 4, 5])
+@pytest.mark.parametrize("middle_qubit", [0, 1, 2, 3, 4])
+@pytest.mark.parametrize("depth", [2, 10])
+@pytest.mark.parametrize("measurements", [True, False])
+def test_star_fix_connectivity(nqubits, depth, middle_qubit, measurements):
+    """Checks that the transpiled circuit can be executed and is equivalent to original."""
+    original = generate_random_circuit(nqubits, depth, middle_qubit=middle_qubit)
+    if measurements:
+        original.add(gates.M(0))
+    transpiler = StarConnectivityRouter(middle_qubit=middle_qubit)
+    transpiled, hardware_qubits = transpiler(original)
+    final_state = backend.execute_circuit(transpiled).state()
+    target_state = backend.execute_circuit(original).state()
+    hardware_qubits = list(hardware_qubits.values())
+    target_state = _transpose_qubits(target_state, hardware_qubits)
+    np.testing.assert_allclose(final_state, target_state)
+
+
+@pytest.mark.parametrize("nqubits", [2, 3, 4, 5])
+@pytest.mark.parametrize("middle_qubit", [0, 1, 2, 3, 4])
+@pytest.mark.parametrize("unitary_dim", [1, 2])
+@pytest.mark.parametrize("depth", [2, 10])
+def test_fix_connectivity_unitaries(nqubits, unitary_dim, depth, middle_qubit):
+    """Checks that the transpiled circuit can be executed and is equivalent to original
+    when using unitaries."""
+    # find the number of qubits for hardware circuit
+    n_hardware_qubits = max(nqubits, middle_qubit + 1)
+
+    original = Circuit(n_hardware_qubits)
+    pairs = list(itertools.combinations(range(n_hardware_qubits), unitary_dim))
+    for _ in range(depth):
+        qubits = pairs[int(np.random.randint(len(pairs)))]
+        original.add(
+            gates.Unitary(
+                random_unitary(2**unitary_dim, backend=NumpyBackend()), *qubits
+            )
+        )
+
+    transpiler = StarConnectivity(middle_qubit=middle_qubit)
+    transpiled, hardware_qubits = transpiler(original)
+    # check that execution results agree with original (using simulation)
+    backend = NumpyBackend()
+    final_state = backend.execute_circuit(transpiled).state()
+    target_state = backend.execute_circuit(original).state()
+    hardware_qubits = list(hardware_qubits.values())
+    target_state = _transpose_qubits(target_state, hardware_qubits)
+    np.testing.assert_allclose(final_state, target_state)
